// src/models/sportradar/basketball/stat-type.ts
export enum StatType {
    Fieldgoal = "fieldgoal",
    Assist = "assist",
    Block = "block",
    AttemptBlocked = "attemptblocked",
    Rebound = "rebound",
    FreeThrow = "freethrow",
    Turnover = "turnover",
    Steal = "steal",
    PersonalFoul = "personalfoul",
    FlagrantFoul = "flagrantfoul",
    TechnicalFoul = "technicalfoul",
    Ejection = "ejection",
    FoulDrawn = "fouldrawn",
    OffensiveFoul = "offensivefoul"
}