// src/models/sportradar/basketball/qualifier.ts
export enum Qualifier {
    SecondChance = "2ndchance",
    FastBreak = "fastbreak",
    FromTurnover = "fromturnover",
    DefensiveGoalTending = "defensivegoaltending",
    PointsInThePaint = "pointsinthepaint",
    Team = "team",
    DeadBall = "deadball",
    OneFreeThrow = "1freethrow",
    TwoFreeThrow = "2freethrow",
    Mandatory = "mandatory",
}